# market_app_api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Market App",
        "description": "Market App - Flutter & NodeJs",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:4000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all Categories",
                "parameters": [
                    {
                        "name": "productName",
                        "in": "query",
                        "required": false,
                        "description": "Search by Product Name",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ProductResponse"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create Product API",
                "parameters": [
                    {
                        "name": "productName",
                        "in": "formData",
                        "description": "Product Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "productDescription",
                        "in": "formData",
                        "description": "Product Description",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "productPrice",
                        "in": "formData",
                        "description": "Product Price",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "productImage",
                        "in": "formData",
                        "description": "Product Image",
                        "required": false,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ProductResponse"
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "Product Id",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Product Details based on Product Id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ProductResponse"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update Product API",
                "parameters": [
                    {
                        "name": "productName",
                        "in": "fomData",
                        "description": "Product Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "productDescription",
                        "in": "formData",
                        "description": "Product Description",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "productPrice",
                        "in": "formData",
                        "description": "Product Price",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "productImage",
                        "in": "formData",
                        "description": "Product Image",
                        "required": false,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ProductResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete Product API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ProductResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Product": {
            "required": [
                "productName",
                "_id",
                "productPrice"
            ],
            "properties": {
                "_id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "productName": {
                    "type": "string"
                },
                "productDescription": {
                    "type": "string"
                },
                "productPrice": {
                    "type": "integer"
                },
                "productImage": {
                    "type": "string"
                }
            }
        },
...
        "ProductResponse": {
            "properties": {
                "message": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Product"
                    }
                }
            }
        },
...
    }
}